ReVamp to new Ethereum ring setup
Blatantly copied from https://github.com/vertigobr/ethereum
This is far better and avoid a lot of the problems with the old one.
The main improvement is that the nodes actually talk to each other.
The internal RPC works well.

Though it is a messy set of shell scripts, it is actually much simpler
in functioning than the old setup